refactor : all headers have been moved to the entry point & others minors changes

+++ public/index.php
@@ -10,7 +10,6 @@ use Model\SessionManager;
 
 $sessionRepository = new SessionManager();
 $sessionRepository->startSession();
-
 use Config\DatabaseConnection;
 
 use Exceptions\UsernameErrorEmptyException;
@@ -42,10 +41,11 @@ use Controller\ArticleController;
 use Controller\UserController;
 use Controller\DownloadController;
 use Controller\HomepageFormController;
-
-
-
-
+use Exceptions\EmailUnavailableException;
+use Exceptions\EmailUnexistException;
+use Exceptions\FormMessageNotSentException;
+use Exceptions\PasswordIncorrectException;
+use Exceptions\UsernameUnavailableException;
 
 $action = "";
 $selection = "";
@@ -88,30 +88,50 @@ if (isset($_GET['action'])) {
         case "sign_up":
             try {
                 $template = "sign_up.twig";
-
-                $paramaters["message"] = $userController->signUpValidator(
+                $signUp = $userController->signUpValidator(
                     $_POST['username'],
                     $_FILES['profile_image'],
                     $_POST["mail"],
                     $_POST["password"]
                 );
+
+                $paramaters["message"] = $signUp;
+
+                if (is_array($signUp)) {
+                    header("HTTP/1.1 302");
+                    header("Location: index.php?selection=blog");
+                }
             } catch (UsernameErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["username_exception"] =
                     UsernameErrorEmptyException::USERNAME_MESSAGE_ERROR_EMPTY;
             } catch (UsernameWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["username_exception"] = UsernameWrongFormatException::USERNAME_MESSAGE_ERROR_WRONG_FORMAT;
             } catch (FileErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["file_exception"] = FileErrorEmptyException::FILE_MESSAGE_ERROR_NO_FILE_SELECTED;
             } catch (FileTypeException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["file_exception"] = FileTypeException::FILE_MESSAGE_ERROR_TYPE_FILE;
             } catch (EmailErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["email_exception"] = EmailErrorEmptyException::EMAIL_MESSAGE_ERROR_EMPTY;
             } catch (EmailWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["email_exception"] = EmailWrongFormatException::EMAIL_MESSAGE_ERROR_WRONG_FORMAT;
             } catch (PasswordErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["password_exception"] = PasswordErrorEmptyException::PASSWORD_MESSAGE_ERROR_EMPTY;
             } catch (PasswordWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["password_exception"] = PasswordWrongFormatException::PASSWORD_MESSAGE_ERROR_WRONG_FORMAT;
+            } catch (UsernameUnavailableException $e) {
+                header("HTTP/1.1 400");
+                $paramaters["username_exception"] = UsernameUnavailableException::USERNAME_UNAVAILABLE_MESSAGE_ERROR . ' ' . $_POST["username"] . " n'est pas disponible ";
+            } catch (EmailUnavailableException $e) {
+                header("HTTP/1.1 400");
+                $paramaters["email_exception"] = EmailUnavailableException::EMAIL_UNAVAILABLE_MESSAGE_ERROR . ' ' . $_POST["mail"] . " n'est pas disponible";
             }
 
             break;
@@ -126,19 +146,27 @@ if (isset($_GET['action'])) {
                     $_SESSION["username"] = $login["username"];
                     $_SESSION["type_user"] = $login["type_user"];
                     $userController->handleInsertSessionData($_SESSION);
+                    header("HTTP/1.1 302");
+                    header("Location: index.php?selection=blog");
                 }
-            }
-            catch (EmailErrorEmptyException $e) {
+            } catch (EmailErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["email_exception"] = EmailErrorEmptyException::EMAIL_MESSAGE_ERROR_EMPTY;
-            }
-            catch (EmailWrongFormatException $e) {
+            } catch (EmailWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["email_exception"] = EmailWrongFormatException::EMAIL_MESSAGE_ERROR_WRONG_FORMAT;
-            }
-            catch (PasswordErrorEmptyException $e) {
+            } catch (PasswordErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["password_exception"] = PasswordErrorEmptyException::PASSWORD_MESSAGE_ERROR_EMPTY;
-            }
-            catch (PasswordWrongFormatException $e) {
+            } catch (PasswordWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["password_exception"] = PasswordWrongFormatException::PASSWORD_MESSAGE_ERROR_WRONG_FORMAT;
+            } catch (EmailUnexistException $e) {
+                header("HTTP/1.1 400");
+                $paramaters["email_exception"] = EmailUnexistException::EMAIL_UNEXIST_MESSAGE_ERROR;
+            } catch (PasswordIncorrectException $e) {
+                header("HTTP/1.1 401");
+                $paramaters["password_exception"] = PasswordIncorrectException::PASSWORD_INCORRECT_MESSAGE_ERROR;
             }
 
             break;
@@ -146,6 +174,17 @@ if (isset($_GET['action'])) {
         case "download_file":
             $template = "homepage.twig";
             $downloadController->handleDownloadFile();
+            if (is_array($downloadController->handleDownloadFile()) && array_key_exists("file_logs", $downloadController->handleDownloadFile())) {
+                header("Content-Length: " . $downloadController->handleDownloadFile()["file_logs"]);
+                header('Content-Description: File Transfer');
+                header("Content-Type: application/pdf");
+                header("Pragma: public");
+                header("Content-Disposition:attachment;filename=cv.pdf");
+                header("HTTP/1.1 200");
+                readfile($downloadController->handleDownloadFile()["file_logs"]);
+            } elseif (is_array($downloadController->handleDownloadFile()) && array_key_exists("code_error", $downloadController->handleDownloadFile())) {
+                header("Location: index.php?action=error&code=" . $downloadController->handleDownloadFile()["code_error"]);
+            }
             break;
 
         case "contact":
@@ -153,25 +192,39 @@ if (isset($_GET['action'])) {
                 $template = "homepage.twig";
                 $paramaters["message"] = $formController->homepageFormValidator($_POST["firstname"], $_POST["lastname"], $_POST["mail"], $_POST["subject"], $_POST["message"]);
             } catch (FirstNameErrorEmptyException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["firstname_exception"] = FirstNameErrorEmptyException::FIRSTNAME_MESSAGE_ERROR_EMPTY;
             } catch (FirstNameWrongFormatException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["firstname_exception"] = FirstNameWrongFormatException::FIRSTNAME_MESSAGE_ERROR_WRONG_FORMAT;
             } catch (LastnameErrorEmptyException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["lastname_exception"] = LastnameErrorEmptyException::LASTNAME_MESSAGE_ERROR_EMPTY;
             } catch (LastnameWrongFormatException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["lastname_exception"] = LastnameWrongFormatException::LASTNAME_MESSAGE_ERROR_WRONG_FORMAT;
             } catch (EmailErrorEmptyException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["email_exception"] = EmailErrorEmptyException::EMAIL_MESSAGE_ERROR_EMPTY;
             } catch (EmailWrongFormatException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["email_exception"] = EmailWrongFormatException::EMAIL_MESSAGE_ERROR_WRONG_FORMAT;
             } catch (SubjectErrorEmptyException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["subject_exception"] = SubjectErrorEmptyException::SUBJECT_MESSAGE_ERROR_EMPTY;
             } catch (SubjectWrongFormatException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["subject_exception"] = SubjectWrongFormatException::SUBJECT_MESSAGE_ERROR_MIN_20_CHARS_MAX_100_CHARS;
             } catch (ContentMessageErrorEmptyException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["content_message_exception"] = ContentMessageErrorEmptyException::CONTENT_MESSAGE_ERROR_EMPTY;
             } catch (ContentMessageWrongFormatException $e) {
+                header('HTTP/1.1 400');
                 $paramaters["content_message_exception"] = ContentMessageWrongFormatException::CONTENT_MESSAGE_ERROR_MIN_20_CHARS_MAX_500_CHARS;
+            } catch (FormMessageNotSentException $e) {
+                header("HTTP/1.1 500");
+                header("Location: index.php?action=error&code=500");
+                $paramaters["message_not_sent_exception"] = FormMessageNotSentException::MESSAGE_SENT_FAILED;
             }
             break;
 
@@ -185,37 +238,49 @@ if (isset($_GET['action'])) {
         case "add_article":
             $template = "admin_add_article.twig";
             try {
-
                 $paramaters = [
                     "article" => $articleController->handleCreateArticleValidator($_POST["title"], $_FILES["image_file"], $_POST["short-phrase"], $_POST["content"], $_POST["tags"], $_SESSION),
                     "session" => $_SESSION,
-
                 ];
+
+                if (is_array($articleController->handleCreateArticleValidator($_POST["title"], $_FILES["image_file"], $_POST["short-phrase"], $_POST["content"], $_POST["tags"], $_SESSION))) {
+                    header("HTTP/1.1 302");
+                    header("Location: index.php?selection=admin_panel");
+                }
             } catch (TitleErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["title_exception"] = TitleErrorEmptyException::TITLE_MESSAGE_ERROR_EMPTY;
             } catch (TitleWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["title_exception"] = TitleWrongFormatException::TITLE_MESSAGE_ERROR_MAX_500_CHARS;
             } catch (FileErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["file_exception"] = FileErrorEmptyException::FILE_MESSAGE_ERROR_NO_FILE_SELECTED;
             } catch (FileTypeException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["file_exception"] = FileTypeException::FILE_MESSAGE_ERROR_TYPE_FILE;
             } catch (ShortPhraseErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["short_phrase_exception"] = ShortPhraseErrorEmptyException::SHORT_PHRASE_MESSAGE_ERROR_EMPTY;
             } catch (ShortPhraseWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["short_phrase_exception"] = ShortPhraseWrongFormatException::SHORT_PHRASE_MESSAGE_ERROR_MAX_500_CHARS;
             } catch (ContentArticleErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["content_article_exception"] = ContentArticleErrorEmptyException::CONTENT_ARTICLE_MESSAGE_ERROR_EMPTY;
             } catch (ContentArticleWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["content_article_exception"] = ContentArticleWrongFormatException::CONTENT_ARTICLE_MESSAGE_ERROR_MAX_5000_CHARS;
             } catch (TagsErrorEmptyException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["tags_exception"] = TagsErrorEmptyException::TAGS_ERROR_EMPTY;
             } catch (TagsWrongFormatException $e) {
+                header("HTTP/1.1 400");
                 $paramaters["tags_exception"] = TagsWrongFormatException::TAGS_MESSAGE_ERROR_MAX_3_TAGS;
             }
             break;
 
         case "update_article":
-
 
             $defautValuesInEachField =
                 [
@@ -231,17 +296,30 @@ if (isset($_GET['action'])) {
             $paramaters = [
                 "message" => $articleController->handleUpdateArticleValidator($_POST["title"], $_FILES["image_file"], $_POST["original_file_path"], $_POST["short_phrase"], $_POST["content"], $_POST["tags"], $_SESSION, $_POST["id_article"]),
                 "default_value" => $defautValuesInEachField,
-
             ];
+
+            if (is_array($articleController->handleUpdateArticleValidator($_POST["title"], $_FILES["image_file"], $_POST["original_file_path"], $_POST["short_phrase"], $_POST["content"], $_POST["tags"], $_SESSION, $_POST["id_article"]))) {
+                header("HTTP/1.1 302");
+                header("Location: index.php?selection=admin_panel");
+            }
 
 
             break;
         case "delete_article":
-            $articleController->handleDeleteArticle($_GET["id"], $_SESSION);
+            if (is_array($articleController->handleDeleteArticle($_GET["id"], $_SESSION))) {
+                header("HTTP/1.1 302");
+                header("Location: index.php?selection=admin_panel");
+                $articleController->handleDeleteArticle($_GET["id"], $_SESSION);
+            }
+
             break;
         case "logout":
-            $logout = $userController->handleLogout($_SESSION);
-            if (is_array($logout) && array_key_exists("logout", $logout) && $logout["logout"] == 1) $sessionRepository->destroySession();
+
+            if (is_array($userController->handleLogout($_SESSION))) {
+                header("HTTP/1.1 302");
+                header("Location: index.php?selection=blog");
+                $sessionRepository->destroySession();
+            }
             break;
     }
 } elseif ((isset($_GET['selection']))) {
